Extract conversation window rendering core
